+++ packages/xcm-core/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * as acc from './account';
 export * as addr from './address';
 export * as big from './big';
+export * as multiloc from './multilocation';